예약 등록 보완 / 행사 등록 보완 / 권한별 메뉴, 드롭다운 메뉴 구현

+++ AMS/src/main/java/com/cafe24/guribyn/event/EventController.java
@@ -44,7 +44,7 @@ public class EventController {
 	// 기본 행사 fx 중복불가
 	@ResponseBody
 	@RequestMapping(value = "eventFx")
-	public String eventFx(@RequestParam ("eventFx") int fx) {
+	public String eventFx(@RequestParam ("eventFx") double fx) {
 		return eventService.eventFx(fx);
 	}
